Create working Tic-Tac-Toe
I wrote a first working version of the game.
I created a class Game, which includes a hash of available fields 1-9,
an array which holds arrays of all combinations of fields to win, a
variable to check which player's turn it is as well as a variable to
check if the game is over
It contains a method to print the current board showing the available
and already chosen fields
To access the variables, I used attr accessor and initialized a new
global object game = Game.new

I created a class Player which asks for a name and a single sign to use
as a symbol representing the player and saves both.
The class has two methods, one to pick a field on the board and sets the
correspinding field as well as the array-value to the players symbol
Then it sets the variable for the current player to the other player
If the user-input is invalid or the field has already been chosen, the
player is told to choose a different field

Once all the variables have been updated according to the player's
choice, another method checks whether any of the winning lines contains
three of the player's symbols
If true, this player is called the winner, else the game continues

This is a first try and all the variables are pretty public, so I will
try to update the code so as little as possible is publicly available